WIP: Setup database syncing
I've started working on the daemon to periodically sync
data pulled from ESPN's API. As of now, migrations are able
to run, but the logic behind updating/creating new team records
needs to be improved. I've also got ranks able to be saved, but
I haven't actually setup teams records yet, so the data is
meaningless.